# artifacts/api-server/src/routes/payments.ts
import { Router, type IRouter } from "express";
import { db, bookingsTable } from "@workspace/db";
import { eq } from "drizzle-orm";
import { CreatePaymentCheckoutBody, CreatePaymentCheckoutResponse } from "@workspace/api-zod";
import { logger } from "../lib/logger";

const router: IRouter = Router();

router.post("/payments/create-checkout", async (req, res): Promise<void> => {
  const parsed = CreatePaymentCheckoutBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { bookingId, depositAmount } = parsed.data;

  // Check booking exists
  const [booking] = await db.select().from(bookingsTable).where(eq(bookingsTable.id, bookingId));
  if (!booking) {
    res.status(404).json({ error: "Reserva no encontrada" });
    return;
  }

  // If Stripe is not configured, return a mock checkout URL for now
  const stripeSecretKey = process.env.STRIPE_SECRET_KEY;
  if (!stripeSecretKey) {
    req.log.warn("Stripe not configured — returning mock checkout");
    // Return a mock response so the frontend can still be tested
    const mockSessionId = `mock_session_${bookingId}_${Date.now()}`;
    await db.update(bookingsTable).set({ stripeSessionId: mockSessionId }).where(eq(bookingsTable.id, bookingId));
    res.json(CreatePaymentCheckoutResponse.parse({
      checkoutUrl: `${req.protocol}://${req.get("host")}/reservar/confirmacion?booking_id=${bookingId}&session_id=${mockSessionId}`,
      sessionId: mockSessionId,
    }));
    return;
  }

  try {
    const stripe = (await import("stripe")).default;
    const stripeClient = new stripe(stripeSecretKey);

    const domains = process.env.REPLIT_DOMAINS?.split(",") ?? [];
    const baseUrl = domains.length > 0 ? `https://${domains[0]}` : `${req.protocol}://${req.get("host")}`;

    const session = await stripeClient.checkout.sessions.create({
      payment_method_types: ["card"],
      line_items: [
        {
          price_data: {
            currency: "eur",
            product_data: {
              name: `Señal de reserva — Cita #${bookingId}`,
              description: "Señal para confirmar su cita de psicología con Alba García Santillana",
            },
            unit_amount: Math.round(depositAmount * 100),
          },
          quantity: 1,
        },
      ],
      mode: "payment",
      success_url: `${baseUrl}/reservar/confirmacion?booking_id=${bookingId}&session_id={CHECKOUT_SESSION_ID}`,
      cancel_url: `${baseUrl}/reservar?cancelled=true`,
      metadata: { bookingId: String(bookingId) },
    });

    await db.update(bookingsTable).set({ stripeSessionId: session.id }).where(eq(bookingsTable.id, bookingId));

    res.json(CreatePaymentCheckoutResponse.parse({
      checkoutUrl: session.url!,
      sessionId: session.id,
    }));
  } catch (err) {
    req.log.error({ err }, "Error creating Stripe checkout session");
    res.status(500).json({ error: "Error al crear la sesión de pago" });
  }
});

router.post("/payments/webhook", async (req, res): Promise<void> => {
  const stripeSecretKey = process.env.STRIPE_SECRET_KEY;
  const webhookSecret = process.env.STRIPE_WEBHOOK_SECRET;

  if (!stripeSecretKey || !webhookSecret) {
    res.sendStatus(200);
    return;
  }

  try {
    const stripe = (await import("stripe")).default;
    const stripeClient = new stripe(stripeSecretKey);
    const sig = req.headers["stripe-signature"] as string;
    const event = stripeClient.webhooks.constructEvent(req.body, sig, webhookSecret);

    if (event.type === "checkout.session.completed") {
      const session = event.data.object as { metadata?: { bookingId?: string } };
      const bookingId = session.metadata?.bookingId ? parseInt(session.metadata.bookingId, 10) : null;
      if (bookingId) {
        await db.update(bookingsTable)
          .set({ depositPaid: true, status: "confirmed" })
          .where(eq(bookingsTable.id, bookingId));
        req.log.info({ bookingId }, "Booking confirmed after payment");
      }
    }

    res.sendStatus(200);
  } catch (err) {
    req.log.error({ err }, "Webhook error");
    res.status(400).json({ error: "Webhook error" });
  }
});

export default router;