# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API for Alba García Santillana - Psicología
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  /** Duration in minutes */
  duration: number;
  /** Full session price in EUR */
  price: number;
  /** Required deposit amount in EUR */
  depositAmount: number;
  isOnline?: boolean;
  isInPerson?: boolean;
}

export interface TimeSlot {
  /** Date in YYYY-MM-DD */
  date: string;
  /** Time in HH:MM */
  time: string;
  available: boolean;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export interface Booking {
  id: number;
  clientName: string;
  clientEmail: string;
  clientPhone: string;
  serviceId: number;
  /** @nullable */
  serviceName?: string | null;
  appointmentDate: string;
  appointmentTime: string;
  /** @nullable */
  notes?: string | null;
  status: BookingStatus;
  depositPaid: boolean;
  /** @nullable */
  depositAmount?: number | null;
  /** @nullable */
  stripeSessionId?: string | null;
  /** @nullable */
  calendarEventId?: string | null;
  createdAt: string;
}

export interface BookingInput {
  /** @minLength 2 */
  clientName: string;
  clientEmail: string;
  clientPhone: string;
  serviceId: number;
  /** Date in YYYY-MM-DD */
  appointmentDate: string;
  /** Time in HH:MM */
  appointmentTime: string;
  notes?: string;
}

export type BookingStatusUpdateStatus = typeof BookingStatusUpdateStatus[keyof typeof BookingStatusUpdateStatus];


export const BookingStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
}

export interface BookingStats {
  total: number;
  pending: number;
  confirmed: number;
  completed: number;
  cancelled: number;
  depositsPaid: number;
}

export interface Review {
  id: number;
  authorName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment: string;
  approved: boolean;
  createdAt: string;
}

export interface ReviewInput {
  /** @minLength 2 */
  authorName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @minLength 10 */
  comment: string;
}

export type ReviewStatsRatingDistribution = {
  '1'?: number;
  '2'?: number;
  '3'?: number;
  '4'?: number;
  '5'?: number;
};

export interface ReviewStats {
  averageRating: number;
  totalReviews: number;
  ratingDistribution: ReviewStatsRatingDistribution;
}

export interface PaymentCheckoutInput {
  bookingId: number;
  depositAmount: number;
}

export interface PaymentCheckoutResult {
  checkoutUrl: string;
  sessionId: string;
}

export type GetAvailabilityParams = {
/**
 * Month in YYYY-MM format
 */
month?: string;
};