# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for Alba García Santillana - Psicología
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings (admin)
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Invalid request
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a specific booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /payments/create-checkout:
    post:
      operationId: createPaymentCheckout
      tags: [payments]
      summary: Create a Stripe checkout session for the booking deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCheckoutInput"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentCheckoutResult"
...
components:
  schemas:
...
    Booking:
      type: object
      required: [id, clientName, clientEmail, clientPhone, serviceId, appointmentDate, appointmentTime, status, depositPaid, createdAt]
      properties:
        id:
          type: integer
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: ["string", "null"]
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        depositPaid:
          type: boolean
        depositAmount:
          type: ["number", "null"]
        stripeSessionId:
          type: ["string", "null"]
        calendarEventId:
          type: ["string", "null"]
        createdAt:
          type: string
    BookingInput:
      type: object
      required: [clientName, clientEmail, clientPhone, serviceId, appointmentDate, appointmentTime]
      properties:
        clientName:
          type: string
          minLength: 2
        clientEmail:
          type: string
          format: email
        clientPhone:
          type: string
        serviceId:
          type: integer
        appointmentDate:
          type: string
          description: "Date in YYYY-MM-DD"
        appointmentTime:
          type: string
          description: "Time in HH:MM"
        notes:
          type: string
    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
...
    PaymentCheckoutInput:
      type: object
      required: [bookingId, depositAmount]
      properties:
        bookingId:
          type: integer
        depositAmount:
          type: number
    PaymentCheckoutResult:
      type: object
      required: [checkoutUrl, sessionId]
      properties:
        checkoutUrl:
          type: string
        sessionId:
          type: string