# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API for Alba García Santillana - Psicología
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  clientName: string;
  clientEmail: string;
  clientPhone: string;
  serviceId: number;
  /** @nullable */
  serviceName?: string | null;
  appointmentDate: string;
  appointmentTime: string;
  /** @nullable */
  notes?: string | null;
  status: BookingStatus;
  depositPaid: boolean;
  /** @nullable */
  depositAmount?: number | null;
  /** @nullable */
  stripeSessionId?: string | null;
  /** @nullable */
  calendarEventId?: string | null;
  createdAt: string;
}