# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API for Alba García Santillana - Psicología
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  /** @minLength 2 */
  clientName: string;
  clientEmail: string;
  clientPhone: string;
  serviceId: number;
  /** Date in YYYY-MM-DD */
  appointmentDate: string;
  /** Time in HH:MM */
  appointmentTime: string;
  notes?: string;
}