# lib/api-zod/src/generated/types/paymentCheckoutInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API for Alba García Santillana - Psicología
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentCheckoutInput {
  bookingId: number;
  depositAmount: number;
}